+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.113.0
 
 - `[options]` `Alias`, `Conflicts`, and `Bound` now supports string slices
+- `[options]` Improved string representation format of `Map`, `V` and option name
 
 ### 12.112.1